Optimize folder send and remove Windows workaround
We were using strings.TrimPrefix() to remove the prefix, but that checks if it contains the prefix. We know that all files within a folder will contain it. It also works on Windows becuase URI uses / as separator and filepath.Walk calls with \ (e.g. "C:/home/dir" == "C:\home\dir").