Remove unnecessary uri code in receive and fix path bug
A uri string was created and sent back. That string "file://path/to/somefile.txt" was then created with "file://" added again.
This also made it so that the fileicon would get an incorrect uri and thus never be able to show the correct folder icon (might also require a change in fyne due to a second bug).
We can stop creating the uri twice and just return the path directly, thus leading to a slight performance improvement and a bug fix.