Rework how settings are fetched on startup
This changes how we store the settings and most importantly how we read them on startup.
The old way of doing it had the undesired consequence that it also saved the preferences
to disk after having loaded them. With that fixed, we should now avoid multiple
unnecessary disk writes on startup. The application should
thus start faster and hopefully avoid other issues in the process.

It also switched some fo the values to reading booleans instead. Should be faster but is a breaking change.

Hopefully a step in the right direction for #22.